refactor(api): split room gateway from sync engine
Move Durable Object lookup and room forwarding into a route-owned gateway so sync-engine.ts only owns HTTP sync response mechanics.

Remove the unused subject parsing path from RoomContext so the Durable Object no longer depends on the host room-name convention.
