+++ apps/api/src/sync-engine.ts
@@ -1,5 +1,4 @@
 import { MAX_PAYLOAD_BYTES } from './constants';
-import type { DispatchRpcRequest, Room } from './room';
 
 type RoomSyncBackend = {
 	sync(
@@ -11,28 +10,6 @@ type RoomSyncBackend = {
 	}>;
 	getDoc(roomName: string): Promise<{ data: Uint8Array; storageBytes: number }>;
 };
-
-export function cloudflareDurableObjectRooms(
-	roomNamespace: DurableObjectNamespace<Room>,
-) {
-	const getRoom = (roomName: string) =>
-		roomNamespace.get(roomNamespace.idFromName(roomName));
-
-	return {
-		handleWebSocket(roomName: string, request: Request) {
-			return getRoom(roomName).fetch(request);
-		},
-		sync(roomName: string, body: Uint8Array) {
-			return getRoom(roomName).sync(body);
-		},
-		getDoc(roomName: string) {
-			return getRoom(roomName).getDoc();
-		},
-		dispatch(roomName: string, request: DispatchRpcRequest) {
-			return getRoom(roomName).dispatch(request);
-		},
-	};
-}
 
 export function createSyncEngine(
 	rooms: RoomSyncBackend,
